feat(detection/configs): moved geoip_db_yaml data to static variable #879
